rename : features 하위 폴더명, 파일명 소문자 및 kebab-case로 변경
