--- src/features/Update_Study_Image/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './Update_study_Image';
