+++ src/features/update-goal-title/ui/index.ts
@@ -3,4 +3,4 @@
 // export * from './Navigation';
 // export * from './MenuItems';
 
-export * from './Update_Study_Info';
+export * from './update-goal-title';
