+++ src/features/update-study-image/api/index.ts
