+++ src/features/update-study-image/api/updateStudyImage.ts
