+++ src/features/update-study-image/index.ts
