+++ src/features/update-study-image/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './update_study_image';
